+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Made auto-correct much less likely to crash due to conflicting corrections ("clobbering"). ([@jonas054][])
 * [#565](https://github.com/bbatsov/rubocop/pull/565): `$GLOBAL_VAR from English library` should no longer be inserted when autocorrecting short-form global variables like `$!`. ([@nevir][])
 * [#566](https://github.com/bbatsov/rubocop/pull/566): Methods that just assign a splat to an ivar are no longer considered trivial writers. ([@nevir][])
+* [#585](https://github.com/bbatsov/rubocop/pull/585): `MethodCallParentheses` should allow methods starting with uppercase letter. ([@bbatsov][])
 
 ## 0.14.1 (10/10/2013)